% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMLogisticLossIterations.R
\name{LMLogisticLossIterations}
\alias{LMLogisticLossIterations}
\title{LMLogisticLossIterations}
\usage{
LMLogisticLossIterations(X.mat, y.vec, max.iterations, step.size)
}
\arguments{
\item{X.mat}{}

\item{y.vec}{}

\item{max.iterations}{}

\item{step.size}{}
}
\description{
LMLogisticLossIterations
}
\examples{
   library(codingProject2)
   data(spam, package = "ElemStatLearn")
   X.mat<-spam[1:10,-58]
   y.vec<-spam[1:10, 58]
   max.iterations <- 50
   step.size <- 0.1
   k <- sigmoid(0)
   res <- LMLogisticLossIterations(X.mat, y.vec, 50, 0.1)
}
