% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMSquareLossEarlyStoppingCV.R
\name{LMSquareLossEarlyStoppingCV}
\alias{LMSquareLossEarlyStoppingCV}
\title{Title}
\usage{
LMSquareLossEarlyStoppingCV(X.mat, y.vec, fold.vec, max.iterations)
}
\arguments{
\item{X.mat}{}

\item{y.vec}{}

\item{fold.vec}{}

\item{max.iterations}{}
}
\description{
Title
}
\examples{
   library(codingProject2)
   data(ozone, package = "ElemStatLearn")
   X.mat<-ozone[1:20,-1]
   y.vec<-ozone[1:20, 1]
   max.iterations <- 30
   step.size <- 0.1
   fold.vec <- sample(rep(1:5, l=nrow(X.mat)))
   res <- LMSquareLossIterations(X.mat, y.vec, fold.vec, max.iterations)
}
