% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMSquareLossIterations.R
\name{LMSquareLossIterations}
\alias{LMSquareLossIterations}
\title{Title}
\usage{
LMSquareLossIterations(X.mat, y.vec, max.iterations, step.size)
}
\arguments{
\item{X.mat}{}

\item{y.vec}{}

\item{max.iterations}{}

\item{step.size}{}
}
\description{
Title
}
\examples{
   library(codingProject2)
   data(ozone, package = "ElemStatLearn")
   X.mat<-ozone[1:10,-1]
   y.vec<-ozone[1:10, 1]
   max.iterations <- 50
   step.size <- 0.1
   res <- LMSquareLossIterations(X.mat, y.vec, 50, 0.1)
   
}
